scrapers/county_permits/requirements.txt
requests
beautifulsoup4
psycopg2-binary
pdfminer.six
